toml

+++ run_dds.sh
@@ -59,17 +59,18 @@ QUECTEL_USB_CTL=$(cat /tmp/usb_quectel_ctl)
 
 
 # detect cell shield SIM ID and write it to file
-_pb "DDS query Quectel cell shield for SIM ID on $QUECTEL_USB_CTL"
-rm "$LI_FILE_ICCID" > /dev/null 2>&1
-echo -ne "AT+QCCID\r" > $QUECTEL_USB_CTL && \
-sleep 0.1 && timeout 1 cat -v < $QUECTEL_USB_CTL | grep QCCID > "$LI_FILE_ICCID"
-
+if [ "${QUECTEL_USB_CTL}" ]; then
+    _pb "DDS query Quectel cell shield for SIM ID on $QUECTEL_USB_CTL"
+    rm "$LI_FILE_ICCID" > /dev/null 2>&1
+    echo -ne "AT+QCCID\r" > $QUECTEL_USB_CTL && \
+    sleep 0.1 && timeout 1 cat -v < $QUECTEL_USB_CTL | grep QCCID > "$LI_FILE_ICCID"
+fi
 
 
 _pb "tweak the connection supervision timeout parameter"
 touch /tmp/200
-sudo cp /tmp/200 /sys/kernel/debug/bluetooth/hci0/supervision_timeout
-sudo cp /tmp/200 /sys/kernel/debug/bluetooth/hci1/supervision_timeout
+sudo cp /tmp/200 /sys/kernel/debug/bluetooth/hci0/supervision_timeout 2> /dev/null
+sudo cp /tmp/200 /sys/kernel/debug/bluetooth/hci1/supervision_timeout 2> /dev/null
